+++ utilities/findmex/findmex.f90
@@ -21,6 +21,13 @@ program findmex
   double precision,allocatable  :: anum(:),masses(:),  hess(:,:), w(:)
   character*3,allocatable       :: aname(:)
   double precision,allocatable  :: cgeom(:)
+  integer,parameter :: MAXCONS=20
+  double precision  :: shift, gtol, dtol, maxd, scale,cons_val(MAXCONS)
+  integer           :: maxiter,ncons,cons_atm(3,MAXCONS)
+  logical  :: ex
+  integer,parameter :: fid=326
+  NAMELIST /MEXOPT/  shift, gtol, dtol, maxd, maxiter, scale, ncons, cons_atm, cons_val
+
 
   print *," ***************************************** "
   print *," *    findmex.x                           * "
@@ -29,6 +36,34 @@ program findmex
   call initPotential()
   call getInfo(natm,nst)
   call DisableEnergyScaling()
+
+! read input values
+  shift=1d-1
+  gtol =1d-1
+  dtol =1d-3
+  maxd =1d0
+  maxiter = 100
+  scale = 1d0
+  ncons = 0
+  cons_atm=0
+  cons_val=0d0
+  inquire(file="mexopt.in",exist=ex)
+  if(ex)then
+    open(unit=fid,file="mexopt.in",access='sequential',form='formatted',&
+        IOSTAT=ios, position='rewind',action='read',status='old')
+    if(ios.eq.0)then
+      read(unit=fid,nml=MEXOPT)
+      if(ncons>MAXCONS)ncons=MAXCONS
+      if(any(cons_atm(1:2,1:ncons)<1) .or. any(cons_atm(3,1:ncons)<0) .or. &
+         any(cons_atm(1,1:ncons)==cons_atm(2,1:ncons)).or.any(cons_atm(1,1:ncons)==cons_atm(3,1:ncons)).or.&
+         any(cons_atm(2,1:ncons)==cons_atm(3,1:ncons)) .or. any(cons_atm(:,1:ncons)>natm)) then
+            print *,"input atom index for each coordinate constraints:"
+            print "(3I4)",cons_atm(:,1:ncons)
+            stop "invalid atom index"
+      end if
+      close(fid)
+    end if
+  end if
 
 ! allocate arrays
   allocate(masses(natm))
@@ -76,7 +111,7 @@ program findmex
   call analysegeom(natm,cgeom,aname,anum,masses,1.9d0,.true.)
 
   ! search for intersections
-  call findx(natm,nst,cgeom,isurf1,isurf2,99,1d-1,1d0,1d-5)
+  call findx(natm,nst,cgeom,isurf1,isurf2,maxiter,shift,gtol,dtol,maxd,scale,ncons,cons_atm,cons_val)
 
   ! print final geometry information
   print *,"---------------  Final Geometries  ------------------"
@@ -141,35 +176,37 @@ end subroutine calcHess
 !This is then used to construct the Hessian and gradient of the Lagrangian.
 !Newton-Raphson procedure is used to find the critical point.
 !Energy part of the Hessian is done numerically
-subroutine findx(natoms,nstate,cgeom,surf1,surf2,maxiter,shift,Etol,Stol)
+subroutine findx(natoms,nstate,cgeom,surf1,surf2,maxiter,shift,Etol,Stol,maxd,scale,nc,cons_atm,cons_val)
   implicit none
-  integer, intent(in)                                 ::  natoms,surf1,surf2,maxiter,nstate
+  integer, intent(in)                                 ::  natoms,surf1,surf2,maxiter,nstate,nc,cons_atm(3,nc)
   double precision,dimension(3*natoms),intent(inout)  ::  cgeom
-  double precision,intent(in)                         ::  shift,Etol,Stol
+  double precision,intent(in)                         ::  shift,Etol,Stol,maxd,scale,cons_val(nc)
 
   integer   :: ndeg  ! number of degenerate states
   integer   :: noffd ! number of off-diagonals that has to vanish 
   integer   :: nlag  ! number of Lagrange multipliers 
   integer   :: neq   ! total number of equations 
   real*8    ::  h(nstate,nstate),cg(3*natoms,nstate,nstate),dcg(3*natoms,nstate,nstate),e(nstate)
-  double precision,dimension(3*natoms)           :: grad !gradient of average energy
-  double precision,dimension(:,:),allocatable    :: hess     ! full hessian of lagrangian 
-  double precision,dimension(:,:),allocatable    :: B    ! Coordinate-LM cross block and its self product 
-  double precision,dimension(:),allocatable :: WORK
-  double precision,dimension(:),allocatable :: lam, dlag    !lagrange multipliers and gradient of lagrangian
-  double precision,dimension(:),allocatable :: w,b1,b2      ! evals of hessian and intermediate vectors 
-  integer,dimension(:),allocatable :: IWORK
-  integer           :: LIWORK, LWORK, itmp(1),INFO  
-  integer  ::  iter  , i,j,lindex
-  double precision            :: nrmG, nrmD,tmp(1)
-  double precision, external  :: dnrm2
-  double precision,  parameter  :: amu2au=1.822888484514D3,au2cm1=219474.6305d0
-  double precision, parameter   :: MAXD = 1D-2
+  double precision,dimension(3*natoms)          :: grad !gradient of average energy
+  double precision,dimension(:,:),allocatable   :: hess     ! full hessian of lagrangian
+  double precision,dimension(:,:),allocatable   :: B    ! Coordinate-LM cross block and its self product
+  double precision,dimension(:),allocatable     :: WORK
+  double precision,dimension(:),allocatable     :: coord_val
+  double precision,dimension(:),allocatable     :: lam, dlag    !lagrange multipliers and gradient of lagrangian
+  double precision,dimension(:),allocatable     :: w,b1,b2      ! evals of hessian and intermediate vectors
+  integer,dimension(:),allocatable  :: IWORK
+  double precision                  :: nrmG, nrmD,tmp(1),bval(12)
+  double precision, external        :: dnrm2
+  double precision, parameter       :: amu2au=1.822888484514D3,au2cm1=219474.6305d0
+  integer           :: offs
+  integer           :: LIWORK, LWORK, itmp(1),INFO
+  integer           ::  iter  , i,j,lindex
 
+  allocate(coord_val(nc))
   ! allocate arrays
   ndeg = surf2-surf1+1
   noffd= ndeg*(ndeg-1)/2
-  nlag = ndeg-1+noffd
+  nlag = ndeg-1+noffd+nc
   neq  = 3*natoms+nlag
   allocate(hess(neq,neq))
   allocate(lam(nlag))
@@ -194,10 +231,8 @@ subroutine findx(natoms,nstate,cgeom,surf1,surf2,maxiter,shift,Etol,Stol)
   do iter=1,maxiter
      call EvaluateSurfgen(cgeom,e,cg,h,dcg,.false.)
      print "(A,I5)","Iteration ",iter
-     print *,"Energies:   "
-     print "(10F20.4)",e*au2cm1 
-     print *,"Lagrange Multipliers:   "
-     print "(10F20.4)",lam
+     print "(A,20F20.4)","Energies : ",e*au2cm1 
+     print "(A,20F20.4)","Lagrange Multipliers : ",lam
      ! use h and cg to store partially diagonalized representation and their gradients
      h = 0d0
      do i=1,nstate
@@ -222,6 +257,26 @@ subroutine findx(natoms,nstate,cgeom,surf1,surf2,maxiter,shift,Etol,Stol)
          lindex=lindex+1
        end do
      end do
+     ! part 3 : coordinate constraint part
+     do i=1,nc
+        hess(lindex,:) =  0d0
+        if(cons_atm(3,i).ne.0)then
+        ! it is an angle
+            call bend(natoms,cons_atm(:,i),cgeom,coord_val(i),bval,5.72957795131d1,0,0d0) ! convert to degrees
+            do j=1,3
+                offs = 3*(cons_atm(j,i)-1)
+                hess(lindex,offs+1:offs+3) =  bval(3*j-2:3*j)
+            end do!j=1,3
+        else !it is a distance
+            call stre(natoms,cons_atm(1,i),cons_atm(2,i),cgeom,coord_val(i),bval)
+            do j=1,2
+                offs = 3*(cons_atm(j,i)-1)
+                hess(lindex,offs+1:offs+3) =  bval(3*j-2:3*j)
+            end do!j=1,2
+        end if
+        hess(1:3*natoms,lindex)=hess(lindex,1:3*natoms)
+        lindex=lindex+1
+     end do!i=1,nc
      !the LM-LM block is zero
      hess(3*natoms+1:neq,3*natoms+1:neq)=0d0
      ! get average gradient
@@ -229,7 +284,7 @@ subroutine findx(natoms,nstate,cgeom,surf1,surf2,maxiter,shift,Etol,Stol)
      do i=surf1,surf2
         dlag(1:3*natoms) = dlag(1:3*natoms)+cg(:,i,i)
      end do
-     dlag=dlag/(surf2-surf1+1)
+     dlag=dlag/(surf2-surf1+1)*scale
      ! calculate optimized lagrange multipliers and update lagrangian gradient
      B = hess(1:3*natoms,3*natoms+1:neq)
      call updateLag(B,3*natoms,nlag,lam,dlag)
@@ -247,6 +302,11 @@ subroutine findx(natoms,nstate,cgeom,surf1,surf2,maxiter,shift,Etol,Stol)
          lindex=lindex+1
        end do
      end do
+     ! part 3: coordinate constraint part
+     do i=1,nc
+        dlag(1:3*natoms)=dlag(1:3*natoms)+lam(lindex)*hess(lindex,1:3*natoms)
+        lindex=lindex+1
+     end do
      ! calculate gradient of lagrangian: lagrange multipliers part
      lindex=3*natoms+1
      do i=surf1+1,surf2
@@ -259,7 +319,12 @@ subroutine findx(natoms,nstate,cgeom,surf1,surf2,maxiter,shift,Etol,Stol)
          dlag(lindex)=h(i,j)
          lindex=lindex+1
        end do
-     end do     
+     end do
+     print "(A,15E15.5)","Coordinate values - constraints :",coord_val-cons_val
+     do i=1,nc
+        dlag(lindex)=coord_val(i)-cons_val(i)
+        lindex=lindex+1
+     end do
      do i=1,3*natoms
        hess(i,i)=hess(i,i)+shift
      end do
@@ -308,7 +373,7 @@ subroutine updateLag(b,nvibs,nlag,lam,grad)
   double precision :: btb(nlag,nlag),v(nlag),dl(nlag),w(nlag)
   integer::LWORK 
   double precision :: WORK(3*nlag*nlag+20)
-  integer :: INFO,ipiv(nlag)
+  integer :: INFO,ipiv(nlag),i
   LWORK=3*nlag*nlag+20
   !Bt.(g+B.deltaLam)=0
   !BtB.deltaLam=-Bt.g
@@ -318,6 +383,13 @@ subroutine updateLag(b,nvibs,nlag,lam,grad)
   call dgemv('T',nvibs,nlag,-1d0,B,nvibs,grad,1,0d0,v,1)
 
   call dsysv('U',nlag,1,BtB,nlag,ipiv,v,nlag,WORK,LWORK,INFO)
-  if(INFO/=0)STOP "UpdateLag: DSYSV Failed."
+  if(INFO/=0)then
+    print *,"INFO=",INFO
+    print *,"Bmatrix:"
+    do i=1,nvibs
+      print "(30F15.7)",b(i,:)
+    end do
+    STOP "updateLag: DSYSV Failed."
+  end if
   lam = v 
 end subroutine updateLag